[FD-414] refactor: GoogleAuthProperty 클래스 명과 패키지 변경
