+++ back-end/src/test/java/com/feedhanjum/back_end/test/config/GoogleAuthMockConfig.java
@@ -1,6 +1,6 @@
 package com.feedhanjum.back_end.test.config;
 
-import com.feedhanjum.back_end.auth.config.GoogleAuthProperty;
+import com.feedhanjum.back_end.auth.config.property.GoogleAuthProperties;
 import com.feedhanjum.back_end.auth.service.GoogleAuthService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -10,7 +10,7 @@ import static org.mockito.Mockito.mock;
 @Configuration
 public class GoogleAuthMockConfig {
     @Bean
-    public GoogleAuthProperty mockGoogleAuthProperty() {
+    public GoogleAuthProperties mockGoogleAuthProperty() {
         return mock();
     }
